pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interrogation-agent"
version = "0.1.0"
description = "LangGraph-based stakeholder hypothesis interrogation service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.129.0,<1.0.0",
  "uvicorn[standard]>=0.40.0,<1.0.0",
  "langgraph>=1.0.8,<2.0.0",
  "langsmith>=0.3.0,<1.0.0",
  "langchain-openai>=1.1.9,<2.0.0",
  "pydantic>=2.12.5,<3.0.0",
  "python-dotenv>=1.2.1,<2.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
